feat: refactored support for BIP322
